refactor: moved hooks to shared folder
