+++ src/components/UI/dropDown/index.tsx
@@ -3,7 +3,7 @@
 import { useRef } from "react";
 
 import { ArrowIcon } from "@/components/icons/svgIcons";
-import { useToggleMenu } from "@/hooks/useToggleMenu";
+import { useToggleMenu } from "@/shared/hooks/useToggleMenu";
 import { cn } from "@/shared/utils/styling";
 import { TDropDown } from "@/types/uiElements";
